epilepsy2bids_requirements.txt
# Core dependencies for epilepsy2bids
epilepsy2bids>=0.1.0
mne>=1.0.0
mne-bids>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
python-dateutil>=2.8.2

# Additional helpful packages
nibabel>=3.2.0
scipy>=1.7.0